+++ frontend/Idle/src/pages/trimPage.jsx
@@ -1,5 +1,4 @@
 import { useContext, useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
 import { styled } from "styled-components";
 import { preloadContext } from "../store/context"
 import { getWithoutQueryAPI } from "../utils/api"
@@ -9,22 +8,17 @@ import TrimMain from "../components/trimPage/trimMain/TrimMain"
 import NormalTrimBoxContainer from "../components/trimPage/trimSub/NormalTrimBoxContainer"
 import FindTrim from "../components/trimPage/findTrim/FindTrim"
 import FindTrimButton from "../components/trimPage/trimSub/FindTrimButton"
-import BlueButton from "../components/common/buttons/BlueButton"
 import FindTrimTooltip from "../components/common/toolTips/FindTrimTooltip"
-import palette from "../styles/palette"
+import TrimConfirmContainer from "../components/trimPage/trimSub/TrimConfirmContainer";
 
 let cachedTrimData = null;
 
 function TrimPage() {
-  const navigate = useNavigate();
   const [toolTipStatus, setToolTipStatus] = useState(true);
   const [trimData, setTrimData] = useState(cachedTrimData);
   const [modalVisible, setModalVisible] = useState(false);
   const { setPreLoadData, preloadImages } = useContext(preloadContext);
 
-  function nextBTNClicked() {
-    navigate("/detail/engines");
-  }
   useEffect(() => {
     getWithoutQueryAPI(PATH.TRIM).then((result) => {
       setTrimData(result);
@@ -41,7 +35,6 @@ function TrimPage() {
     });
   }, []);
 
-
   function findButtonClicked() {
     setToolTipStatus(false);
     setModalVisible(true);
@@ -52,13 +45,7 @@ function TrimPage() {
       <StWrapper>
         <StBottomContainer>
           {trimData ? <NormalTrimBoxContainer data={trimData} onMouseEnter={preloadImages} /> : <Loading />}
-          <StConfirmContainer>
-            <StConfirmHeader>
-              <Title>트림 선택</Title>
-              <Description>원하는 트림을 선택해주세요.</Description>
-            </StConfirmHeader>
-            <BlueButton text={"다음"} onClick={nextBTNClicked} />
-          </StConfirmContainer>
+          <TrimConfirmContainer />
         </StBottomContainer>
         <FindTrimButton onClick={findButtonClicked} onMouseEnter={preloadImages} />
         <TrimSelectContainer onClick={() => { setToolTipStatus(false); }}>
@@ -91,45 +78,12 @@ const StBottomContainer = styled.div`
   left: 128px;
 `;
 
-const StConfirmContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 78px;
-`;
-
-const StConfirmHeader = styled.div`
-  display: inline-flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 4px;
-`;
-
 const TrimSelectContainer = styled.div`
   display: flex;
   position: absolute;
   bottom: 10px;
   left: 55%;
   gap: 15px;
-`;
-
-const Title = styled.h1`
-  color: ${palette.Black};
-  font-family: "Hyundai Sans Text KR";
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 24px;
-  letter-spacing: -0.48px;
-`;
-const Description = styled.p`
-  color: ${palette.Black};
-  font-family: "Hyundai Sans Text KR";
-  font-size: 13px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 165%;
-  letter-spacing: -0.39px;
 `;
 
 const StTooltipContainer = styled.div`
